% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{EmbedInTree}
\alias{EmbedInTree}
\title{Embed data in a single tree created from GetTrees}
\usage{
EmbedInTree(x, tree)
}
\arguments{
\item{x}{input data (matrix, data.frame, or data.table). Shape is n x p.}

\item{tree}{processed data.table output from xgboost::xgb.model.dt.tree}
}
\value{
binary embedding matrix. Of shape n x #-terminal-nodes-in-tree
}
\description{
Send a datapoint through a decision tree, embedding the data in binary tree-leaf space, that tree
being a structure created with the GetTrees function.
}
