% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{Fbboost}
\alias{Fbboost}
\title{Fit a penalized least squares model to data in high-dimensional tree space}
\usage{
Fbboost(x, y, booster = NULL, crossVal = TRUE, nFolds = 5, ...)
}
\arguments{
\item{x}{input data (matrix, data.frame, or data.table). Shape is n x p.}

\item{y}{input labels or name of target variable present in x.}

\item{booster}{a fitted xgb.Booster model if you've already trained the booster. Can also supply
parameters to train an xgboost model here}

\item{crossVal}{should linear model be cross validated?}

\item{nFolds}{if crossVal is TRUE, number of folds used to train linear model}

\item{...}{arguments to xgboost::xgb.train}
}
\value{
fitted glmnet model
}
\description{
Fit a penalized least squares model to data in high-dimensional tree space using a
boosted tree model
}
