% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{GetTreeTable}
\alias{GetTreeTable}
\title{Back out trees from boosted tree model from xgboost package}
\usage{
GetTreeTable(featureNames, model)
}
\arguments{
\item{featureNames}{names of data as it was sent into xgboost::xgb.train}

\item{model}{an xgb.Booster model}
}
\value{
a data.table of trees
}
\description{
Make the output of xgboost::xgb.model.dt.tree more useful
}
