% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{NormalizedEntropy}
\alias{NormalizedEntropy}
\title{Normalized entropy calculator}
\usage{
NormalizedEntropy(trueVec, predVec)
}
\arguments{
\item{trueVec}{numeric vector of true class labels}

\item{predVec}{numeric vector of class probability prediction}
}
\value{
numeric value
}
\description{
Calculate normalized entropy, i.e. binary
cross entropy divided by entropy of background probability of positive class label;
defined in He, Pan, et al 2014
}
