% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{NormalizedEntropySummary}
\alias{NormalizedEntropySummary}
\title{Calculate Normalized Entropy across samples}
\usage{
NormalizedEntropySummary(data, lev = GetDefaultLevels()[2], model = NULL)
}
\arguments{
\item{data}{data.table provided by caret::train}

\item{lev}{str positive class label}

\item{model}{NULL required by caret::train}
}
\value{
named list
}
\description{
summary function to use Normalized Entropy metric in caret::train
see ?caret::defaultSummary
}
