% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ValidateXgboostInput}
\alias{ValidateXgboostInput}
\title{Validate input passed to xgb.train}
\usage{
ValidateXgboostInput(inputNames)
}
\arguments{
\item{inputNames}{names of params being sent as kwargs to do.call(xgboost, ...)}
}
\description{
Make sure we're passing in valid input into xgb.train
}
