% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{WrapXgbTrain}
\alias{WrapXgbTrain}
\title{xgboost::xgb.train wrapper}
\usage{
WrapXgbTrain(x, y, ...)
}
\arguments{
\item{x}{input data (matrix, data.frame, or data.table). Shape is n x p.}

\item{y}{input labels or name of target variable present in x.}

\item{...}{arguments to xgboost::xgb.train}
}
\value{
xgb.Booster model
}
\description{
Simple wrapper for xgboost::xgb.train
}
